+++ CHANGELOG.md
@@ -45,6 +45,8 @@ The changes are relative to the previous release, unless the baseline is specifi
   Header OBU, for compatibility with libraries that wrongly ignore the colr box.
 * Use a "quality to quantizer (QP)" mapping formula designed for AOM_TUNE_IQ.
 * Set tuning before applying the user-provided specific aom codec options.
+* Use AOM_TUNE_PSNR by default when encoding alpha with libaom because
+  AOM_TUNE_SSIM causes ringing for alpha.
 
 ### Removed since 1.3.0